Enable authorization - new way

+++ app/src/files/files.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import {createHash} from 'crypto';
 import { ObjectID } from 'bson';
@@ -7,6 +7,8 @@ import { extname } from 'path';
 import { unlink, readFile } from 'fs/promises';
 import * as dotenv from 'dotenv'
 import { PaginationDto } from '../universal-dto/pagination.dto';
+import { RolesService } from 'src/roles/roles.service';
+import { Privilege } from '@prisma/client';
 const env = dotenv.config({path: '.env'}).parsed;
 
 const number_of_list = (+env.PAGINATION_ELEMENTS_PER_PAGE)
@@ -14,10 +16,12 @@ const number_of_list = (+env.PAGINATION_ELEMENTS_PER_PAGE)
 @Injectable()
 export class FilesService {
     constructor(
-        private prisma : PrismaService
+        private prisma : PrismaService,
+        private roles_service : RolesService
     ){}
     
     async public_file_list(user_id: string, query : PaginationDto){
+        if(!await this.roles_service.checkPrivilages(user_id, [Privilege.FILES_CAN_READ])){throw new ForbiddenException();}
         return await this.prisma.file.findMany({skip: query.offset*number_of_list, take:number_of_list, include:{
             owners:{
                 select:{
@@ -40,11 +44,18 @@ export class FilesService {
     async public_downloadFileById(file_id : string, user_id: string, res : any){
         if(!ObjectID.isValid(file_id)){throw new NotFoundException()}
         const file_entity = await this.prisma.file.findUnique({where:{id:file_id}});
+        if(!file_entity.owners_ids.includes(user_id)){
+            if(!await this.roles_service.checkPrivilages(user_id, [Privilege.FILES_CAN_READ])){throw new ForbiddenException();}
+        }
         const file_data = createReadStream(file_entity.path);
         file_data.pipe(res);
     }
 
     async public_uploadFile(file : Express.Multer.File, user_id : string){
+        if(!await this.roles_service.checkPrivilages(user_id, [Privilege.FILES_CAN_READ, Privilege.FILES_CAN_WRITE])){
+            await unlink(file.path);
+            throw new ForbiddenException();
+        }
         let ext = extname(file.originalname);
         let size = file.size;
         const file_buffer = await readFile(file.path);
@@ -70,6 +81,9 @@ export class FilesService {
         if(!ObjectID.isValid(file_id)){throw new NotFoundException()}
         let file = await this.prisma.file.findUnique({where: {id: file_id}});
         if(!file){throw new NotFoundException()}
+        if(!file.owners_ids.includes(user_id)){
+            if(!await this.roles_service.checkPrivilages(user_id, [Privilege.FILES_CAN_READ, Privilege.FILES_CAN_DELETE])){throw new ForbiddenException();}
+        }
         file = await this.prisma.file.update({where: {id:file_id}, data:{owners:{disconnect:{id: user_id}}}});
         if(file.owners_ids.length === 0){
             await unlink(file.path);
